modify the waf build/extract logic to use base64 encoding for the binary data
- storing real binary data with the --tools option would crash
a lot. This switches the logic to use a base64 encoded blob written
out as comments at the end of the file. This does increase the size
of waf, but it's much safer and allows you to bundle binary objects
like whl files for python utilities, etc. with less hiccups