setup refactors and new sdists command
- change the setup command to use pre/pip/post hooks, where pre is
for system provisioning, pip is only for pip installations using the
pyd member's api, and post is for anything else. This allows the
sdists command to sync down the sdists using the pip hook, which
simplifies the configuration of the project a lot.
- the sdists command will sync down the pip files you install via
the pyd.pyenv_install calls in the pip-hook.